Add the ga and idm plugins
GA was added because it it loads IDM.
